+++ src/leaf/bridge.ts
@@ -7,7 +7,7 @@ import { DelayedPromise } from '../commons/promise';
 import { newIdGenerator } from '../commons/utils';
 
 /**
- * Available leaf interface commands
+ * Available leaf bridge commands
  */
 export const enum LeafBridgeCommands {
     Info = "info",
@@ -26,7 +26,13 @@ export interface LeafBridgeElement {
 }
 
 /**
- * LeafInterface read and write to the IDE dedicated python leaf's extension
+ * This custom error is thrown when we are in the 'profile out of sync' state
+ * where envvars cannot be computed by leaf
+ */
+export const PROFILE_OUT_OF_SYNC_ERROR = new Error("Profile out of sync");
+
+/**
+ * LeafBridge read and write to the IDE dedicated python leaf's extension
  */
 export class LeafBridge {
 
@@ -36,7 +42,7 @@ export class LeafBridge {
     private pendingRequests: { [key: string]: DelayedPromise<LeafBridgeElement | undefined> } = {}; // pending promises callbacks
 
     public constructor() {
-        // Launch interface
+        // Launch bridge
         let pathToExec = join(__filename, '..', '..', '..', 'python-src', 'leaf-bridge.py');
 
         // Copy system env vars
@@ -54,17 +60,17 @@ export class LeafBridge {
         });
 
         // Listen from stdout
-        this.process.stdout.addListener("data", (chunk: Buffer | string) => this.onBridgeResponse(chunk));
+        this.process.stdout.addListener("data", (chunk: Buffer | string) => this.onBridgeData(chunk));
 
         // Listen from stderr
-        this.process.stderr.addListener("data", (chunk: Buffer | string) => console.log(`[Leaf Bridge] Error from leaf interface: ${chunk.toString()}`));
+        this.process.stderr.addListener("data", (chunk: Buffer | string) => console.log(`[Leaf Bridge] Error from leaf bridge: ${chunk.toString()}`));
 
         // Configure stdin
         this.process.stdin.setDefaultEncoding('utf-8');
     }
 
     /**
-     * Send request to leaf extension interface
+     * Send request to leaf extension bridge
      */
     public send(cmd: LeafBridgeCommands): Promise<LeafBridgeElement | undefined> {
         let out = new DelayedPromise<LeafBridgeElement | undefined>();
@@ -81,43 +87,59 @@ export class LeafBridge {
     }
 
     /**
-     * Called on interface response.
-     * Concatenate and parse the resolve or reject the corresponding promise
+     * Called on bridge data
+     * Concatenate, parse then resolve or reject the corresponding promise
      */
-    private onBridgeResponse(chunk: Buffer | string) {
+    private onBridgeData(chunk: Buffer | string) {
         this.stdoutBuffer += chunk.toString();
         if (this.stdoutBuffer.endsWith("\n")) {
-            let lines: string[] = this.stdoutBuffer.split(/\r?\n/).filter((value) => value.length > 0);
+            this.stdoutBuffer
+                .split(/\r?\n/) // Split to lines
+                .filter(value => value.length > 0) // Discard empty lines
+                .forEach(line => this.onBridgeResponse(line)); // Parse line
             this.stdoutBuffer = "";
-            for (let line of lines) {
-                try {
-                    let anyResponse: any = JSON.parse(line);
-                    if (this.pendingRequests && anyResponse.id in this.pendingRequests) {
-                        let pendingRequest = this.pendingRequests[anyResponse.id];
-                        if (anyResponse.result) {
-                            pendingRequest.resolve(anyResponse.result);
-                            console.log(`[Leaf Bridge] Response received for id '${anyResponse.id}': '${line.substring(0, 10)}...'`);
-                        } else if (anyResponse.error) {
-                            pendingRequest.resolve(undefined);
-                            console.log(`[Leaf Bridge] Error received for id '${anyResponse.id}': '${anyResponse.error.message}...'`);
-                        } else {
-                            pendingRequest.resolve(undefined);
-                            console.log(`[Leaf Bridge] No 'result' or 'error' child node in parsed json '${line.substring(0, 10)}...'`);
-                        }
-                        delete this.pendingRequests[anyResponse.id];
-                    } else {
-                        console.log(`[Leaf Bridge] Unknown response id ${anyResponse.id}`);
-                        throw new Error(`Unknown response id ${anyResponse.id}`);
-                    }
-                } catch (e) {
-                    console.log(`[Leaf Bridge] ${e}: ${chunk}`);
-                }
-            }
         }
     }
 
+    /**
+     * Called on bridge response
+     * Parse then resolve or reject the corresponding promise
+     */
+    private onBridgeResponse(line: string) {
+        try {
+            let anyResponse: any = JSON.parse(line);
+            if (this.pendingRequests && anyResponse.id in this.pendingRequests) {
+                let pendingRequest = this.pendingRequests[anyResponse.id];
+                if (anyResponse.result) {
+                    pendingRequest.resolve(anyResponse.result);
+                    console.log(`[Leaf Bridge] Response received for id '${anyResponse.id}': '${line.substring(0, 10)}...'`);
+                } else if (anyResponse.error) {
+                    let errorType = anyResponse.error.type;
+                    if (errorType === 'ProfileOutOfSyncException') {
+                        pendingRequest.reject(PROFILE_OUT_OF_SYNC_ERROR);
+                    } else {
+                        pendingRequest.resolve(undefined);
+                    }
+                    console.log(`[Leaf Bridge] Error received for id '${anyResponse.id}': '${errorType}/${anyResponse.error.message}'`);
+                } else {
+                    pendingRequest.resolve(undefined);
+                    console.log(`[Leaf Bridge] No 'result' or 'error' child node in parsed json '${line.substring(0, 10)}...'`);
+                }
+                delete this.pendingRequests[anyResponse.id];
+            } else {
+                console.log(`[Leaf Bridge] Unknown response id ${anyResponse.id}`);
+                throw new Error(`Unknown response id ${anyResponse.id}`);
+            }
+        } catch (e) {
+            console.log(`[Leaf Bridge] ${e}: ${line}`);
+        }
+    }
+
+    /**
+     * Terminate bridge binary and remove all listeners
+     */
     public async dispose() {
-        this.send(LeafBridgeCommands.Exit);
+        await this.send(LeafBridgeCommands.Exit);
         this.process.stdin.end();
         this.process.stdout.removeAllListeners();
         this.process.stderr.removeAllListeners();