+++ src/leaf/profiles.ts
@@ -4,7 +4,12 @@ import { window, StatusBarItem, StatusBarAlignment } from "vscode";
 import { LeafManager, LeafEvent } from './core';
 import { Command, View } from '../commons/identifiers';
 import { ProfileQuickPickItem, ProfileTreeItem, PackageTreeItem } from './uiComponents';
-import { showMultiStepQuickPick, toItems, TreeDataProvider2 } from '../commons/uiUtils';
+import { showMultiStepQuickPick, toItems, TreeDataProvider2, QuickPickItem2 } from '../commons/uiUtils';
+
+// The icons used in status bar
+// Must be taken from the [octicon](https://octicons.github.com) icon set.
+const OK_ICON = '$(check)';
+const WARNING_ICON = '$(alert)';
 
 /**
  * Leaf manager.
@@ -24,7 +29,7 @@ export class LeafProfileStatusBar extends TreeDataProvider2 {
     // So lets add status bar
     this.leafStatusbar = window.createStatusBarItem(StatusBarAlignment.Left, 11);
     this.toDispose(this.leafStatusbar);
-    this.leafStatusbar.text = "Loading current profile...";
+    this.leafStatusbar.text = `${WARNING_ICON} Loading current profile...`;
     this.leafStatusbar.tooltip = "Current Leaf profile";
     this.leafStatusbar.show();
 
@@ -37,13 +42,16 @@ export class LeafProfileStatusBar extends TreeDataProvider2 {
     // Subscribe to leaf events
     this.leafManager.addListener(LeafEvent.CurrentProfileChanged, this.onProfileChanged, this);
     this.leafManager.addListener(LeafEvent.ProfilesChanged, this.refresh, this);
-    this.setInitialState();
+    this.leafManager.addListener(LeafEvent.onProfileOutOfSync, this.refreshStatusbar, this);
+
+    // Set initial state
+    this.refreshStatusbar();
   }
 
   /**
-   * Async initialisation
+   * Refresh the profile name in the status bar
    */
-  private async setInitialState() {
+  private async refreshStatusbar() {
     try {
       await this.onProfileChanged(undefined, await this.leafManager.getCurrentProfileName());
     } catch (reason) {
@@ -57,7 +65,17 @@ export class LeafProfileStatusBar extends TreeDataProvider2 {
    */
   private async switchProfile(): Promise<void> {
     let profiles = await this.leafManager.getProfiles();
-    let items: ProfileQuickPickItem[] = toItems(profiles, ProfileQuickPickItem);
+    let items: QuickPickItem2[] = toItems(profiles, ProfileQuickPickItem);
+    let syncProfileItem: QuickPickItem2 = {
+      id: "cmd",
+      properties: undefined,
+      label: "Fix profile",
+      description: "Execute 'leaf profile sync'",
+      compareTo: () => 0
+    };
+    if (await this.leafManager.isOutOfSync()) {
+      items.push(syncProfileItem);
+    }
     let result = await showMultiStepQuickPick(
       "leaf profile switch",
       undefined,
@@ -65,9 +83,15 @@ export class LeafProfileStatusBar extends TreeDataProvider2 {
       "Please select the profile you want to switch to...",
       items
     );
-    if (result) {
-      return this.leafManager.switchProfile(result.id);
+
+    // User want to sync current profile
+    if (result === syncProfileItem) {
+      return this.leafManager.syncCurrentProfile();
     }
+
+    // User picked a profile to switch to
+    result = result as ProfileQuickPickItem;
+    return this.leafManager.switchProfile(result.id);
   }
 
 	/**
@@ -95,6 +119,10 @@ export class LeafProfileStatusBar extends TreeDataProvider2 {
     }
   }
 
+  /**
+   * Remove a package from its profile
+   * @param packageItem the package item selected in tree view
+   */
   private async removePackage(packageItem: PackageTreeItem | undefined) {
     if (!packageItem || !packageItem.parent) {
       throw new Error('Command not available from the palette; try the Leaf Profiles view');
@@ -105,12 +133,18 @@ export class LeafProfileStatusBar extends TreeDataProvider2 {
   /**
    * Profile has changed, let's update status bar
    */
-  private onProfileChanged(_oldProfileName: string | undefined, newProfileName: string | undefined) {
+  private async onProfileChanged(_oldProfileName: string | undefined, newProfileName: string | undefined) {
     if (this.leafStatusbar) {
-      this.leafStatusbar.text = newProfileName ? newProfileName : 'No profile';
+      let state = newProfileName && !(await this.leafManager.isOutOfSync());
+      let stateIcon = state ? OK_ICON : WARNING_ICON;
+      let profile = newProfileName || 'No profile';
+      this.leafStatusbar.text = `${stateIcon} ${profile}`;
     }
   }
 
+  /**
+   * @returns the list of items that will be shown at the root of the tree view
+   */
   protected async getRootElements(): Promise<ProfileTreeItem[]> {
     let lm = this.leafManager;
     return toItems(await lm.getProfiles(), class extends ProfileTreeItem {